move h2o_socket_is_traced to socket.c
