+++ lib/core/util.c
@@ -945,115 +945,6 @@ void h2o_cleanup_thread(void)
     h2o_mem_clear_recycle(&h2o_socket_buffer_prototype.allocator);
 }
 
-#if H2O_USE_DTRACE && defined(__linux__)
-#include <linux/bpf.h>
-#include <linux/unistd.h>
-#include "h2o-probes.h"
-#include "include/h2o/ebpf.h"
-#include <sys/stat.h>
-
-int h2o_tracing_map_fd = 0;
-uint32_t h2o_tracing_map_last_attempt = 0;
-
-void open_map()
-{
-    // only check every second
-    uint32_t now = (uint32_t)time(NULL);
-    if (h2o_tracing_map_last_attempt == now)
-        return;
-
-    __sync_bool_compare_and_swap(&h2o_tracing_map_last_attempt, h2o_tracing_map_last_attempt, now);
-
-    // check if map exists at path
-    struct stat s;
-    if (stat(&H2O_EBPF_MAP_PATH[0], &s) == -1) {
-        // map path unavailable, cleanup fd if needed and leave
-        if (h2o_tracing_map_fd > 0) {
-            close(h2o_tracing_map_fd);
-            __sync_fetch_and_and(&h2o_tracing_map_fd, 0);
-        }
-        return;
-    }
-
-    if (h2o_tracing_map_fd > 0)
-        return; // map still exists and we have a fd
-
-    union bpf_attr attr;
-    memset(&attr, 0, sizeof(attr));
-    attr.pathname = (uint64_t)(unsigned long)&H2O_EBPF_MAP_PATH[0];
-
-    int fd = syscall(__NR_bpf, BPF_OBJ_GET, &attr, sizeof(attr));
-    if (fd <= 0)
-        return;
-
-    // try set map fd if not set before
-    if (!__sync_bool_compare_and_swap(&h2o_tracing_map_fd, 0, fd))
-        close(fd); // if a fd was set beforehand, don't set the new fd, and close it
-}
-
-inline int lookup_map(const void *key, const void *value)
-{
-    union bpf_attr attr;
-    memset(&attr, 0, sizeof(attr));
-    attr.map_fd = h2o_tracing_map_fd;
-    attr.key = (uint64_t)(unsigned long)key;
-    attr.value = (uint64_t)(unsigned long)value;
-    return syscall(__NR_bpf, BPF_MAP_LOOKUP_ELEM, &attr, sizeof(attr)) == -1 ? -1 : 1;
-}
-
-inline void read_ip_port(struct sockaddr *sa, void *ip, uint16_t *port)
-{
-    if (sa->sa_family == AF_INET) {
-        struct sockaddr_in *sin = (void *)sa;
-        memcpy(ip, &sin->sin_addr, sizeof(sin->sin_addr));
-        *port = sin->sin_port;
-    } else if (sa->sa_family == AF_INET6) {
-        struct sockaddr_in6 *sin = (void *)sa;
-        memcpy(ip, &sin->sin6_addr, sizeof(sin->sin6_addr));
-        *port = sin->sin6_port;
-    }
-}
-
-int h2o_socket_is_traced(h2o_socket_t *sock)
-{
-    if (sock == NULL)
-        return 0;
-
-    if (sock->_is_traced != 0)
-        return sock->_is_traced;
-
-    // try open map if not opened
-    open_map();
-    if (h2o_tracing_map_fd <= 0)
-        return 1; // map can't be opened, fallback accepting probe
-
-    // define key/vals - we are only interrested in presence of the key, discard values
-    struct h2o_ebpf_map_key_t key;
-    memset(&key, 0, sizeof(key));
-    void *vals = NULL;
-
-    // get sock/peer ip/ports
-    struct sockaddr_storage loc, rem;
-    h2o_socket_getsockname(sock, (void *)&loc);
-    h2o_socket_getpeername(sock, (void *)&rem);
-
-    // read ip/ports, put parsed val into key structure
-    read_ip_port((void *)&loc, &key.ipa, &key.porta);
-    read_ip_port((void *)&rem, &key.ipb, &key.portb);
-
-    // read ip family
-    key.family = loc.ss_family == AF_INET6 ? 6 : 4;
-
-    // lookup map for our key
-    return sock->_is_traced = lookup_map(&key, &vals);
-}
-#else
-int h2o_socket_is_traced(h2o_socket_t *sock)
-{
-    return 1;
-}
-#endif
-
 int h2o_conn_is_traced(h2o_conn_t *conn)
 {
     return h2o_socket_is_traced(conn->callbacks->get_socket(conn));
